+++ src/main/java/org/hv/pocket/session/actions/OperateDictionary.java
@@ -128,7 +128,7 @@ public interface OperateDictionary {
     int update(AbstractEntity entity, boolean cascade) throws SQLException, IllegalAccessException;
 
     /**
-     * 删除实体
+     * 级联删除
      *
      * @param entity 实体对象
      * @return 影响行数
@@ -136,4 +136,15 @@ public interface OperateDictionary {
      * @throws IllegalAccessException e
      */
     int delete(AbstractEntity entity) throws SQLException, IllegalAccessException;
+
+    /**
+     * 删除实体
+     *
+     * @param entity  实体对象
+     * @param cascade 是否进行级联更新操作
+     * @return 影响行数
+     * @throws SQLException           语句异常
+     * @throws IllegalAccessException e
+     */
+    int delete(AbstractEntity entity, boolean cascade) throws SQLException, IllegalAccessException;
 }